refactor(ckeditor5/codeblock): split dropdown into own plugin
